Feature: 모달 기본 레이아웃 및 프롭 설정
