+++ src/ui/Modal.module.scss
@@ -1,0 +1,10 @@
+.modal {
+  width: fit-content;
+  height: fit-content;
+  position: relative;
+  border: none;
+  border-radius: 8px;
+  &::backdrop {
+    background-color: rgba(0, 0, 0, 0.5);
+  }
+}
